requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
weasyprint==60.1
pdfkit==1.0.0
pywin32==306; platform_system=="Windows"
jinja2==3.1.2
